+++ BrcmPatchRAM/BrcmFirmwareStore.cpp
@@ -32,6 +32,7 @@
  * Zlib Decompression
  ***************************************/
 #include <libkern/zlib.h>
+#include <libkern/crypto/sha1.h>
 
 extern "C"
 {
@@ -475,7 +476,18 @@ OSArray* BrcmFirmwareStore::loadFirmware(UInt16 vendorId, UInt16 productId, OSSt
         AlwaysLog("Non-compressed firmware.\n");
 
     configuredData->release();
-    
+
+#ifdef DEBUG
+    SHA1_CTX ctx;
+    uint8_t hash[SHA1_RESULTLEN];
+    SHA1Init(&ctx);
+    SHA1Update(&ctx, firmwareData->getBytesNoCopy(), firmwareData->getLength());
+    SHA1Final(hash, &ctx);
+    DebugLog("Firmware SHA1: %02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x\n",
+             hash[0], hash[1], hash[2], hash[3], hash[4], hash[5], hash[6], hash[7], hash[8], hash[9],
+             hash[10], hash[11], hash[12], hash[13], hash[14], hash[15], hash[16], hash[17], hash[18], hash[19]);
+#endif
+
     OSArray* instructions = parseFirmware(firmwareData);
     firmwareData->release();